# api-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Demo classifieds marketplace API (C2C listings: items, categories, favorites, chat, listing promotion).\nNot affiliated with Avito, OLX, or any other classifieds platform.",
        "title": "Classifieds Marketplace API",
        "contact": {},
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "List categories",
                "parameters": [
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "description": "Locale (en or ru, default en)",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_categories.ListResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "categories"
                ],
                "summary": "Get category by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Category UUID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "enum": [
                            "en",
                            "ru"
                        ],
                        "type": "string",
                        "description": "Locale (en or ru, default en)",
                        "name": "locale",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/internal_api_categories.SingleResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "github_com_dionisvl_classifieds_api-go_internal_api.ProblemDetails": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "detail": {
                    "type": "string"
                },
                "status": {
                    "type": "integer"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "internal_api_categories.CategoryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "names": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    }
                },
                "parent_id": {
                    "type": "string"
                },
                "slug": {
                    "type": "string"
                },
                "sort_order": {
                    "type": "integer"
                }
            }
        },
        "internal_api_categories.ListResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/internal_api_categories.CategoryResponse"
                    }
                }
            }
        },
        "internal_api_categories.SingleResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "$ref": "#/definitions/internal_api_categories.CategoryResponse"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}